added alt manager prototype
I have now added an Alt Manager prototype. It will allow us to implement all the required features without the actual design being finished. It's also a bit more user-friendly for people who already use this client. Anyway, it supports Cracked, Microsoft and Altening authentication thanks to the Elixir library. It is also able to fully serialise and deserialise from the accounts.json. There are still a few unimplemented features that are already shown in the UI, but will be added functionally soon. There's almost no response to the user at the moment, so if there's an error, you won't even know, haha...