Fix SilentHotbar not changing mining speed (#1536)
Currently works, but not the cleanest solution because I had to add the canGetSlot function to SilentHotbar. For some weird reason, getMainHandStack gets called before mc.player gets initialized. If someone has a fix for this I would be happy to change it. If not I think it should be fine like this. It at least works now.

closes #1533