fix repeatables not working due to continueLoop being false
sync() has been used before in this case and fixes the problem without doing anything weird. It seems that due to the fact that we initialise the coroutine before we initialise the continueLoop variable, we cause the loop to break. sync() will delay this and therefore fix the problem.

Friendly reminder to everyone that this is why we make release and nightly builds.