An alternative delay implementation for SuperKnockBack. (Like AutoAccount) (#1709)
Decided to replace the current delay method with something with the way AutoAccount module has it. This fixes waiting a tick forcefully for attacking modules that work on the repeatable sequence. (KillAura, Trigger)

I also added an event that checks for module mode change. In SuperKnockBack's case it's useful in case the user switches modes, so the sequence doesn't get stuck forever.