release: 0.1.0
So, we have decided that since we are a long way from an actual 1.0.0 release, we are going to go with a lower version number from now on and go up with actual releases. This means that from now on there will be actual nextgen releases and nextgen will be the default client in both our launcher and our website.

All builds that have been previous called LiquidBounce 1.0.0 are now internally named v0.0.1

This is our FIRST actual release of LiquidBounce Nextgen.