feat: InventoryTracker and Invsee command (#4441)
This module tracks the inventories of other players by storing their held items along with the timestamp. The tracked inventory can then be viewed with the invsee command. I also added two new events: EntityEquipmentChangeEvent, which is called when an EntityEquipmentUpdateS2CPacket is received and LoreQueryEvent, which is called when the tooltip for an item is requested.