feat(KillAura): Target Selection Improvement (#5769)
Drastically improves the way KillAura selects targets when it is unable to break shields, either when no axe is selected or when we do not have Auto Weapon enabled with axe selection.

Often misunderstood. Shield Breaker is INTEGRATED with Auto Weapon. Select Axe for the [Blocked by Shield] option, and it will automatically break shields on attack if configured with at least 1 tick of on-time.