feat(Phase/Blink): automatic phase-in and phase-out  (#6125)
As soon as we start phasing, e.g. walking horizontally through a block or sneaking vertically, the phase will start blinking. As soon as the blocks disappear or turn off the Phase module, we stop phasing and stop blinking. This allows us to blink to a position that was not accessible at the moment we started blinking.

Very similar to [ModuleFreeze]'s BalanceWarp option, where we can buffer ticks and release them once the blocks have disappeared. This mode works on Grim Anti-Cheat thanks to the fact that we flush incoming packets and pass their outgoing packets before flushing the original packets.

This is very useful for every game-mode that keeps players in a box before starting the game.