refactor(SafeWalk): new on edge mode (#6186)
This is a rewrite of the on-edge detection mode, which now checks when moving away from edges. There are also multiple modes to prevent falling, e.g., Stop, Center and Invert, as well as a timer to keep input for longer, and a sneak and jump option.

The Eagle On Ledge option has been removed as it is redundant given the [Eagle] module and settings. Simulate is removed as well (replaced by on edge).