feat: marketplace (#5226)
Allows subscribing to marketplace items such as themes and scripts which are automatically being installed and updated.

Introduces:
- `.marketplace list <type> [featured]`
- `.marketplace subscribe <id>`
- `.marketplace unsubscribe <id>`
- `.marketplace update [id]`
- `.marketplace revisions list <id>`

All item creation and editing commands are currently commented out because it will be easier to maintain them through https://liquidbounce.net/marketplace in the future. This commit also introduces an automatic theme upload workflow to the LiquidBounce Marketplace and improved README for the theme workspace.